+++ src/dx/publish/mod.rs
@@ -15,9 +15,11 @@
 pub use result::{PublishResponseBody, PublishResult};
 pub mod result;
 
+#[doc(inline)]
+pub use builders::PublishMessageBuilder;
+pub mod builders;
+
 use crate::core::headers::{APPLICATION_JSON, CONTENT_TYPE};
-#[cfg(feature = "serde")]
-use crate::providers::deserialization_serde::SerdeDeserializer;
 use crate::{
     core::{
         Deserializer, PubNubError, Serialize, Transport, TransportMethod, TransportRequest,
@@ -25,283 +27,10 @@ use crate::{
     },
     dx::PubNubClient,
 };
-use derive_builder::Builder;
+use builders::{PublishMessageViaChannel, PublishMessageViaChannelBuilder};
+use std::collections::HashMap;
 use std::ops::Not;
-use std::{collections::HashMap, rc::Rc};
 use urlencoding::encode;
-
-/// The [`PublishMessageBuilder`] is used to publish a message to a channel.
-///
-/// This struct is used to publish a message to a channel. It is used by the [`publish_message`] method of the [`PubNubClient`].
-/// The [`publish_message`] method is used to publish a message to a channel.
-/// The [`PublishMessageBuilder`] is used to build the request to be sent to the [`PubNub`] network.
-///
-/// # Examples
-/// ```rust
-/// # use pubnub::{PubNubClientBuilder, Keyset};
-///
-/// # #[tokio::main]
-/// # async fn main() -> Result<(), Box<dyn std::error::Error>> {
-/// let mut pubnub = // PubNubClient
-/// # PubNubClientBuilder::with_reqwest_transport()
-/// #     .with_keyset(Keyset{
-/// #         subscribe_key: "demo",
-/// #         publish_key: Some("demo"),
-/// #         secret_key: None,
-/// #     })
-/// #     .with_user_id("user_id")
-/// #     .build()?;
-///
-/// pubnub.publish_message("hello world!")
-///     .channel("my_channel")
-///     .execute()
-///     .await?;
-///
-/// # Ok(())
-/// # }
-/// ```
-///
-/// [`PublishMessageBuilder`]: crate::dx::publish::PublishMessageBuilder]
-/// [`publish_message`]: crate::dx::PubNubClient::publish_message`
-/// [`PubNubClient`]: crate::dx::PubNubClient
-/// [`PubNub`]:https://www.pubnub.com/
-pub struct PublishMessageBuilder<'pub_nub, T, M>
-where
-    T: Transport,
-    M: Serialize,
-{
-    pub_nub_client: &'pub_nub PubNubClient<T>,
-    message: M,
-    seqn: u16,
-}
-
-impl<'pub_nub, T, M> PublishMessageBuilder<'pub_nub, T, M>
-where
-    T: Transport,
-    M: Serialize,
-{
-    /// The [`channel`] method is used to set the channel to publish the message to.
-    ///
-    /// [`channel`]: crate::dx::publish::PublishMessageBuilder::channel
-    #[cfg(feature = "serde")]
-    pub fn channel<S>(
-        self,
-        channel: S,
-    ) -> PublishMessageViaChannelBuilder<'pub_nub, T, M, SerdeDeserializer>
-    where
-        S: Into<String>,
-    {
-        PublishMessageViaChannelBuilder::<T, M, SerdeDeserializer> {
-            pub_nub_client: Some(self.pub_nub_client),
-            seqn: Some(self.seqn),
-            ..Default::default()
-        }
-        .message(self.message)
-        .channel(channel.into())
-        .deserialize_with(SerdeDeserializer)
-    }
-
-    #[cfg(not(feature = "serde"))]
-    pub fn channel<S>(self, channel: S) -> PublishMessageDeserializerBuilder<'pub_nub, T, M>
-    where
-        S: Into<String>,
-    {
-        PublishMessageDeserializerBuilder {
-            pub_nub_client: self.pub_nub_client,
-            message: self.message,
-            seqn: self.seqn,
-            channel: channel.into(),
-        }
-    }
-}
-
-/// The [`PublishMessageDeserializer`] is used to publish a message to a channel.
-///
-/// This struct is used to publish a message to a channel. It is used by the [`publish_message`] method of the [`PubNubClient`].
-/// It adds the deserializer to the [`PublishMessageBuilder`].
-///
-/// The [`publish_message`] method is used to publish a message to a channel.
-///
-/// See more information in the [`PublishMessageBuilder`] struct and the [`Deserializer`] trait.
-///
-/// # Examples
-/// ```rust
-/// # use pubnub::{PubNubClientBuilder, Keyset};
-/// use pubnub::{
-///     dx::publish::PublishResponse,
-///     core::{Deserializer, PubNubError}
-/// };
-/// # #[tokio::main]
-/// # async fn main() -> Result<(), Box<dyn std::error::Error>> {
-///
-/// struct MyDeserializer;
-///
-/// impl<'de> Deserializer<'de, PublishResponseBody> for MyDeserializer {
-///    fn deserialize(&self, response: &'de [u8]) -> Result<PublishResponse, PubNubError> {
-///    // ...
-///    # Ok(PublishResponse)
-/// }
-///
-///
-/// let mut pubnub = // PubNubClient
-/// # PubNubClientBuilder::with_reqwest_transport()
-/// #     .with_keyset(Keyset{
-/// #         subscribe_key: "demo",
-/// #         publish_key: Some("demo"),
-/// #         secret_key: None,
-/// #     })
-/// #     .with_user_id("user_id")
-/// #     .build()?;
-///
-/// pubnub.publish_message("hello world!")
-///    .channel("my_channel")
-///    .deserialize_with(MyDeserializer)
-///    .execute()
-///    .await?;
-/// # Ok(())
-/// # }
-/// ```
-///
-/// [`PublishMessageDeserializer`]: crate::dx::publish::PublishMessageDeserializer
-/// [`publish_message`]: crate::dx::PubNubClient::publish_message`
-/// [`PubNubClient`]: crate::dx::PubNubClient
-/// [`PublishMessageBuilder`]: crate::dx::publish::PublishMessageBuilder
-/// [`Deserializer`]: crate::core::Deserializer
-#[cfg(not(feature = "serde"))]
-pub struct PublishMessageDeserializerBuilder<'pub_nub, T, M>
-where
-    T: Transport,
-    M: Serialize,
-{
-    pub_nub_client: &'pub_nub PubNubClient<T>,
-    message: M,
-    seqn: u16,
-    channel: String,
-}
-
-#[cfg(not(feature = "serde"))]
-impl<'pub_nub, T, M> PublishMessageDeserializerBuilder<'pub_nub, T, M>
-where
-    T: Transport,
-    M: Serialize,
-{
-    /// The [`deserialize_with`] method is used to set the deserializer to use to deserialize the response.
-    /// It's important to note that the deserializer must implement the [`Deserializer`] trait for
-    /// the [`PublishResponse`] type.
-    ///
-    /// [`deserialize_with`]: crate::dx::publish::PublishMessageDeserializerBuilder::deserialize_with
-    /// [`Deserializer`]: crate::core::Deserializer
-    /// [`PublishResponse`]: crate::core::publish::PublishResponse
-    pub fn deserialize_with<D>(
-        self,
-        deserializer: D,
-    ) -> PublishMessageViaChannelBuilder<'pub_nub, T, M, D>
-    where
-        for<'de> D: Deserializer<'de, PublishResponseBody>,
-    {
-        PublishMessageViaChannelBuilder {
-            pub_nub_client: Some(self.pub_nub_client),
-            seqn: Some(self.seqn),
-            deserializer: Some(Rc::new(deserializer)),
-            ..Default::default()
-        }
-        .message(self.message)
-        .channel(self.channel)
-    }
-}
-
-/// The [`PublishMessageViaChannelBuilder`] is used to publish a message to a channel.
-/// This struct is used to publish a message to a channel. It is used by the [`publish_message`] method of the [`PubNubClient`].
-///
-/// This is next step in the publish process. The [`PublishMessageViaChannelBuilder`] is used to build the request to be sent to the [`PubNub`] network.
-///
-/// # Examples
-/// ```rust
-/// # use pubnub::{PubNubClientBuilder, Keyset};
-///
-/// # #[tokio::main]
-/// # async fn main() -> Result<(), Box<dyn std::error::Error>> {
-/// let mut pubnub = // PubNubClient
-/// # PubNubClientBuilder::with_reqwest_transport()
-/// #     .with_keyset(Keyset{
-/// #         subscribe_key: "demo",
-/// #         publish_key: Some("demo"),
-/// #         secret_key: None,
-/// #     })
-/// #     .with_user_id("user_id")
-/// #     .build()?;
-///
-/// pubnub.publish_message("hello world!")
-///     .channel("my_channel")
-///     .execute()
-///     .await?;
-///
-/// # Ok(())
-/// # }
-/// ```
-///
-/// [`PublishMessageViaChannelBuilder`]: crate::dx::publish::PublishMessageViaChannelBuilder
-/// [`publish_message`]: crate::dx::PubNubClient::publish_message
-/// [`PubNub`]:https://www.pubnub.com/
-/// [`PubNubClient`]: crate::dx::PubNubClient
-#[derive(Builder)]
-#[builder(pattern = "owned", build_fn(private))]
-pub struct PublishMessageViaChannel<'pub_nub, T, M, D>
-where
-    T: Transport,
-    M: Serialize,
-    D: for<'de> Deserializer<'de, PublishResponseBody>,
-{
-    #[builder(setter(custom))]
-    pub_nub_client: &'pub_nub PubNubClient<T>,
-
-    #[builder(setter(custom))]
-    seqn: u16,
-
-    // TODO: Moved to heap to avoid partial move, but this is not ideal. ref[1]
-    #[builder(setter(custom))]
-    deserializer: Rc<D>,
-
-    /// Message to publish
-    message: M,
-
-    /// Channel to publish to
-    #[builder(setter(into))]
-    channel: String,
-
-    /// Switch that decide if the message should be stored in history
-    #[builder(setter(strip_option), default = "None")]
-    store: Option<bool>,
-
-    /// Switch that decide if the transaction should be replicated
-    /// following the PubNub replication rules.
-    ///
-    /// See more at [`PubNub replication rules`]
-    ///
-    /// [`PubNub replication rules`]:https://www.pubnub.com/pricing/transaction-classification/
-    #[builder(default = "true")]
-    replicate: bool,
-
-    /// Set a per-message TTL time to live in storage.
-    #[builder(setter(strip_option), default = "None")]
-    ttl: Option<u32>,
-
-    /// Switch that decide if the message should be published using POST method.
-    #[builder(setter(strip_option), default = "false")]
-    use_post: bool,
-
-    /// Object to send additional information about the message.
-    #[builder(setter(strip_option), default = "None")]
-    meta: Option<HashMap<String, String>>,
-
-    /// Space ID to publish to.
-    #[builder(setter(strip_option), default = "None")]
-    space_id: Option<String>,
-
-    /// Message type to publish.
-    #[builder(setter(strip_option), default = "None")]
-    message_type: Option<String>,
-}
 
 fn bool_to_numeric(value: bool) -> String {
     if value { "1" } else { "0" }.to_string()
@@ -406,32 +135,6 @@ where
     M: Serialize,
     D: for<'de> Deserializer<'de, PublishResponseBody>,
 {
-    /// Deserializer to use to deserialize the response.
-    /// It's important to note that the deserializer must implement the [`Deserializer`] trait for
-    /// the [`PublishResponse`] type.
-    pub fn deserialize_with<D2>(
-        self,
-        deserializer: D2,
-    ) -> PublishMessageViaChannelBuilder<'pub_nub, T, M, D2>
-    where
-        D2: for<'de> Deserializer<'de, PublishResponseBody>,
-    {
-        PublishMessageViaChannelBuilder {
-            pub_nub_client: self.pub_nub_client,
-            channel: self.channel,
-            message: self.message,
-            seqn: self.seqn,
-            store: self.store,
-            replicate: self.replicate,
-            ttl: self.ttl,
-            use_post: self.use_post,
-            meta: self.meta,
-            space_id: self.space_id,
-            message_type: self.message_type,
-            deserializer: Some(Rc::new(deserializer)),
-        }
-    }
-
     /// Execute the request and return the result.
     /// This method is asynchronous and will return a future.
     /// The future will resolve to a [`PublishResponse`] or [`PubNubError`].
